third_party/blink/web_tests/flag-specific/disable-site-isolation-trials/webaudio/internals/audioworkletprocessor-gc.https-expected.txt

This is a testharness.js-based test.
[FAIL] Test GC of AudioWorkletProcessor
  assert_equals: the initial handler count should be zero. expected 0 but got 5
Harness: the test ran to completion.
